backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-extra-fields==3.0.2
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.2
filelock==3.13.4
flake8==7.0.0
gunicorn==21.2.0
identify==2.5.35
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
mccabe==0.7.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pypng==0.20220715.0
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.3
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
sentry-sdk==1.45.0
setuptools==69.2.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
